#394: move the shard worker pool to scripts/ as bench transport
Five review findings have now landed on this pool's lifecycle, all the
same shape: a transition that leaves a promise unresolved or an invariant
unrecorded. Three were fixed; a fourth I found myself while checking one
of my own claims, and a fifth arrived from review. Fixing each one
individually was working, but the rate of arrival is the signal, not the
individual bugs.

The distinction that settles it: **the pool is load-bearing for the
evidence, not the behaviour.** Every timing in 3.5a came through it, so
deleting it would cost reproducibility of 3.26x / 1.15x. But nothing
about the index depends on it -- it is in no export barrel, no production
path reaches it, and the builder's default runner is
`inProcessShardRunner`, which runs shards sequentially in-process with no
worker lifecycle at all. So none of the five findings is reachable from
anything this package exports.

Moving it to `scripts/` costs neither. The harness still runs, and the
lifecycle leaves the shipped surface entirely: `scripts/` is not compiled,
not published (`files` is `compiled/**`), and not matched by Jest.

What ships is the builder and the merge. `ShardRunner` stays in `src/` and
is now documented as what it always was -- the transport contract, with
the Node pool as one bench-grade reference implementation. Anything that
can move byte offsets out and a ShardOutcome back satisfies it; a browser
Worker pool would be another.

Two consequences, stated rather than buried:

- **The pool's 7 tests no longer run.** Jest matches `compiled/**` only,
  so a script cannot be tested by it. That is the intended trade -- bench
  code held to a bench bar -- but it means the three fixed lifecycle bugs
  are now protected by nothing except the header that describes them.
- **Two defects are known and unfixed** in the moved file, listed in its
  header: `terminate()` can resolve while an eviction is still inside
  `spawnWarmWorker`, so it is a teardown request rather than a quiescence
  barrier. Harmless for a benchmark that spawns, runs and exits. Not
  harmless in a load path, which is why whoever promotes this to `src/`
  owns the state-machine pass first.

Verified inert: `data/index.ifc` through the moved pool at N=1 and N=2 is
byte-identical to the serial build, same digest as before the move.
