+++ scripts/README.md
@@ -18,6 +18,7 @@ performance investigation needs already exists here.
 | `agg_pager_trace.mjs` | Every window request the store-backed load makes, tagged by caller (#616): per-phase chunk requests / hits / loads / bytes, a per-chunk load histogram, the load-order shape, and per-wave records for `AggregateExtractPager`. `--structure-only` reports just the aggregate shape (relationship count, targets per relationship, address spans, and the product-worklist / aggregate-target partition that #539 is about) for one open. `--chunk` / `--cap` override the window geometry, which both turns "the window is too small" into an A/B and pins the adaptive cap for a baseline run. Also classifies every load as compulsory or capacity (against a ghost list of recent evictions) and reports per-interval trigger statistics — the D3D-vs-PSB discriminator the #616 policy triggers on — and `--dump-stream` writes the request stream for `pager_policy_sim.mjs` |
 | `agg_closure_probe.mjs` | Where one aggregate target's paging closure actually lives (#616): samples targets across the file, pages each one's extraction closure with a fresh `seen` set, and reports the closure's chunk footprint plus the entity types sitting outside the product's own chunk. This is what shows a D3D target's closure spanning 208 MB of a 213 MB file |
 | `pager_policy_sim.mjs` | Replay a recorded pager request stream through candidate residency policies (#616). Valid because the `WindowedStepBufferProvider` request stream is window-independent — proven identical at every chunk/cap tested — so one instrumented load captures the whole reference string and any LRU policy's *load count* follows offline, exactly. Turns picking a trigger threshold from a 3-minute D3D load per candidate into a sub-second replay. Capture with `agg_pager_trace.mjs --dump-stream`. Models load counts, not time, and replays sequentially (so it is pessimistic about retention) — calibrate the fixed-cap row against the in-vivo number before trusting it |
+| `shard_worker_pool_node.mjs` | **Bench transport** for `sharded_index_report.mjs` (#394): a Node `worker_threads` pool implementing the builder's `ShardRunner` contract. Deliberately not in `src/` — it is unpublished, untested by Jest, and its header lists lifecycle defects that are known and unfixed, because a harness one person runs on a quiet box is held to a lower bar than shipped code. Promoting it means owning that lifecycle first |
 | `sharded_index_report.mjs` | Serial vs sharded index build over the **shipped** builder (#394 M2): warm and cold wall at each N, per-shard times, shard-only efficiency, and a byte-identity gate plus SHA-256 digest against the single-threaded build. Drives `src/step/parsing/sharded_index_builder.ts` through the Node worker pool rather than carrying its own copy of the logic, so its numbers are about the code that ships. The N=1 row forces the shard path (the shipped builder delegates to the serial one there) and is labelled as such |
 | `index_shard_spike.mjs` | The **spike** that preceded the above (#394 M7/M8), kept as the record of how the direction was proven: its own copy of the shard/merge logic, and a `--selftest` that walks adversarial fixtures (quoted `;#…=`, block comments, `''` escapes) at every split offset. For new work use `sharded_index_report.mjs` and `src/step/parsing/sharded_index_builder.test.ts`, which is where the self-test now lives as a real test |
 | `step_nonproduct_survey.py` | Survey of non-product STEP entity usage across the corpus |
